Atomized places and trainer admin pages
