+++ src/atoms/formElements/LabelAdmin.js
@@ -1,0 +1,7 @@
+export function LabelAdmin({ htmlfor = "name", children }) {
+	return (
+		<label htmlFor={htmlfor} className="block text-sm font-medium text-gray-700">
+			{children}
+		</label>
+	);
+}
