+++ src/atoms/index.js
@@ -45,3 +45,11 @@ export { ModalWithClose } from "./ModalWithClose";
 export { CarouselFull } from "./CarouselFull";
 export { CarouselMultiple } from "./CarouselMultiple";
 export { GridGallery } from "./GridGallery";
+export { CardInputWrapper } from "./CardInputWrapper";
+export { LabelAdmin } from "./formElements/LabelAdmin";
+export { InputAdmin } from "./formElements/InputAdmin";
+export { TextAreaAdmin } from "./formElements/TextAreaAdmin";
+export { UploadImage } from "./formElements/UploadImage";
+export { CustomSelect } from "./formElements/CustomSelect";
+export { ContactLabel } from "./formElements/ContactLabel";
+export { ContactInput } from "./formElements/ContactInput";
