+++ src/molecules/formElements/InputAdminWrapper.js
@@ -1,0 +1,10 @@
+import { LabelAdmin, InputAdmin } from "@src/atoms";
+
+export function InputAdminWrapper({ formik, name, type, description }) {
+	return (
+		<>
+			<LabelAdmin htmlfor={name}>{description}</LabelAdmin>
+			<InputAdmin formik={formik} name={name} type={type} />
+		</>
+	);
+}
