+++ src/molecules/index.js
@@ -7,8 +7,8 @@ export { BigCardWithImage } from "./BigCardWithImage";
 export { ReferenceCard } from "./ReferenceCard";
 export { TeamMemberCard } from "./TeamMemberCard";
 export { ReferenceSection } from "./ReferenceSection";
-export { TextAreaWrapper } from "./TextAreaWrapper";
-export { InputWrapper } from "./InputWrapper";
+export { TextAreaWrapper } from "./formElements/TextAreaWrapper";
+export { InputWrapper } from "./formElements/InputWrapper";
 export { FooterLinksWrapper } from "./FooterLinksWrapper";
 export { PlaceCard } from "./PlaceCard";
 export { TrainerCard } from "./TrainerCard";
@@ -25,3 +25,8 @@ export { Sorter } from "./Sorter";
 export { ProfilePopup } from "./ProfilePopup";
 export { ResponsiveGallery } from "./ResponsiveGallery";
 export { ProfileImageLink } from "./ProfileImageLink";
+export { InputAdminWrapper } from "./formElements/InputAdminWrapper";
+export { TextAreaAdminWrapper } from "./formElements/TextAreaAdminWrapper";
+export { UploadImageWrapper } from "./formElements/UploadImageWrapper";
+export { ContactInputWrapper } from "./formElements/ContactInputWrapper";
+export { SelectWrapper } from "./formElements/SelectWrapper";
