+++ src/organisms/index.js
@@ -16,3 +16,5 @@ export { RatingList } from "./RatingList";
 export { ContactBusinessForm } from "./ContactBusinessForm";
 export { ChangePassword } from "./ChangePassword";
 export { ForgotenPasswordModal } from "./ForgotenPasswordModal";
+export { AdminPlaceForm } from "./AdminPlaceForm";
+export { AdminTrainerForm } from "./AdminTrainerForm";
